Only install steam if disk image not present.

+++ scripts/osx-install.sh
@@ -93,16 +93,18 @@ sudo sed -i "" -e "s/timeout 20/timeout 1/" /EFI/refind/refind.conf
 
 # Install steam in a case-insensitive disk image
 # Source: http://blog.andersonshatch.com/2010/05/13/using-steam-on-mac-with-case-sensitive-drive/
-installcask steam
-hdiutil create -size 30G -fs HFS+ -layout NONE -type SPARSEBUNDLE -volname steam ~/steam
-hdiutil mount ./steam.sparsebundle
-cp -av /opt/homebrew-cask/Caskroom/steam/stable/Steam.app /Volumes/steam/
-mkdir /Volumes/steam/steam\ library /Volumes/steam/steam\ content
-ln -s /Volumes/steam/steam\ library ~/Library/Application\ Support/Steam
-sudo ln -s /Volumes/ /volumes
-brew cask uninstall steam
-# TODO: Find a way to create OSX alias to /Applications (to get automount)
-# See: https://en.wikipedia.org/wiki/Alias_(Mac_OS)
+if [! -e "~/steam.sparsebundle" ]; then
+    brew install cask steam
+    hdiutil create -size 30G -fs HFS+ -layout NONE -type SPARSEBUNDLE -volname steam ~/steam
+    hdiutil mount ~/steam.sparsebundle
+    cp -av /opt/homebrew-cask/Caskroom/steam/stable/Steam.app /Volumes/steam/
+    mkdir /Volumes/steam/steam\ library /Volumes/steam/steam\ content
+    ln -s /Volumes/steam/steam\ library ~/Library/Application\ Support/Steam
+    sudo ln -s /Volumes/ /volumes
+    brew cask uninstall steam
+    # TODO: Find a way to create OSX alias to /Applications (to get automount)
+    # See: https://en.wikipedia.org/wiki/Alias_(Mac_OS)
+fi
 
 # Clean things up
 brew linkapps
